healthbackend/requirements.txt
Flask==2.3.3
langchain==0.1.14
langchain-google-genai==0.0.11
google-generativeai
python-dotenv
